GeneticDeployerServer/requirements.txt
connexion >= 2.6.0
connexion[swagger-ui] >= 2.6.0
python_dateutil == 2.8.1
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
jmetalpy==1.5.4
numpy==1.23.1
nose2==0.9.1
PyYAML==5.1.2
tabulate==0.8.10
tensorboard==2.9.0
tensorboardX==2.5.1
pandas==1.5.1
